chapter8:validation

+++ internal/models/snippets.go
@@ -18,7 +18,7 @@ type SnippetModel struct {
 	DB *sql.DB
 }
 
-func (m *SnippetModel) Insert(title string, content string, expires uint) (uint, error) {
+func (m *SnippetModel) Insert(title string, content string, expires uint16) (uint, error) {
 	stmt := `INSERT INTO snippets (title, content, created, expires)
     VALUES(?, ?, UTC_TIMESTAMP(), DATE_ADD(UTC_TIMESTAMP(), INTERVAL ? DAY))`
